[REF] account: merge res_bank & res_partner_bank
Two files were used to define the same class ResPartnerBank
There was no point in keeping both
ResBank is also a class that exists, so we keep res_partner_bank.py

While at it, cleaned up the code a bit
