+++ addons/account/models/__init__.py
@@ -30,6 +30,5 @@ from . import digest
 from . import res_users
 from . import ir_actions_report
 from . import res_currency
-from . import res_bank
 from . import mail_thread
 from . import account_report
